pyproject.toml
[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "hp_transfer_benchmarks"
version = "0.1.0"
description = ""
authors = ["Danny Stoll, Diane Wagner"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
ConfigSpace = "^0.4.11"
openml = "^0.10.2"
more_itertools = "^8.3.0"
scipy = "^1.4.1"
scikit-learn = "0.19.2"
h5py = "^2.0.0"
hpolib2 = {git = "https://github.com/automl/HPOlib2.git", rev = "v.0.0.4"}

[tool.poetry.dev-dependencies]
jupyter = "^1.0"
pre-commit = "^1.20"
black = "^19.10b0"
isort = "^4.3"
pylint = {git = "https://github.com/PyCQA/pylint.git", rev = "3649312233b39a74b4f631bae1e242f5bc7bde2c"}

[tool.black]
line-length = 90
target-version = ['py37']

[tool.isort]
line_length = 90
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap=0
use_parentheses = true
force_single_line = true
atomic = true
lines_after_imports = 2
lines_between_types = 1
not_skip = "__init__.py"
filter_files = true

[tool.pylint.messages_control]
disable = """
all
"""
enable = """
# ------------------------------------
# Spelling
# ------------------------------------
invalid-characters-in-docstring,wrong-spelling-in-comment,wrong-spelling-in-docstring,
# ------------------------------------
# Basic checks
# ------------------------------------
not-in-loop,function-redefined,continue-in-finally,abstract-class-instantiated,star-needs-assignment-target,
duplicate-argument-name,return-in-init,too-many-star-expressions,nonlocal-and-global,return-outside-function,
return-arg-in-generator,invalid-star-assignment-target,bad-reversed-sequence,nonexistent-operator,
yield-outside-function,init-is-generator,nonlocal-without-binding,lost-exception,assert-on-tuple,
dangerous-default-value,duplicate-key,useless-else-on-loop,expression-not-assigned,confusing-with-statement,
unnecessary-lambda,pointless-statement,unnecessary-pass,unreachable,eval-used,exec-used,using-constant-test,
deprecated-lambda,blacklisted-name,misplaced-comparison-constant,singleton-comparison,unneeded-not,
consider-iterating-dictionary,consider-using-enumerate,empty-docstring,unidiomatic-typecheck,
# ------------------------------------
# Async
# ------------------------------------
not-async-context-manager,yield-inside-async-function,
# ------------------------------------
# Typecheck
# ------------------------------------
invalid-unary-operand-type,unsupported-binary-operation,not-callable,redundant-keyword-arg,assignment-from-no-return,
assignment-from-none,not-context-manager,repeated-keyword,missing-kwoa,no-value-for-parameter,invalid-sequence-index,
invalid-slice-index,unexpected-keyword-arg,unsupported-membership-test,unsubscriptable-object,
# ------------------------------------
# Exceptions
# ------------------------------------
bad-except-order,catching-non-exception,bad-exception-context,notimplemented-raised,raising-bad-type,
raising-non-exception,misplaced-bare-raise,duplicate-except,nonstandard-exception,binary-op-exception,
bare-except,
# ------------------------------------
# Stdlib
# ------------------------------------
bad-open-mode,redundant-unittest-assert,boolean-datetime,deprecated-method
# ------------------------------------
# Imports
# ------------------------------------
import-error,import-self,reimported,relative-import,deprecated-module,wildcard-import,misplaced-future,cyclic-import,
wrong-import-position,ungrouped-imports,multiple-imports,
# ------------------------------------
# Variables
# ------------------------------------
unpacking-non-sequence,invalid-all-object,unbalanced-tuple-unpacking,undefined-variable,undefined-all-variable,
used-before-assignment,cell-var-from-loop,global-variable-undefined,redefined-builtin,redefine-in-handler,
unused-import,unused-argument,unused-wildcard-import,unused-variable,global-variable-not-assigned,
undefined-loop-variable,global-statement,global-at-module-level,
# ------------------------------------
# Strings
# ------------------------------------
format-needs-mapping,truncated-format-string,missing-format-string-key,mixed-format-string,too-few-format-args,
bad-str-strip-call,too-many-format-args,bad-format-character,format-combined-specification,bad-format-string-key,
bad-format-string,missing-format-attribute,missing-format-argument-key,unused-format-string-argument,
unused-format-string-key,invalid-format-index,
# ------------------------------------
# String Constant
# ------------------------------------
anomalous-unicode-escape-in-string,anomalous-backslash-in-string,
# ------------------------------------
# Elif
# ------------------------------------
simplifiable-if-statement,
# ------------------------------------
# Logging
# ------------------------------------
logging-format-truncated,logging-too-few-args,logging-too-many-args,logging-unsupported-format,
# ------------------------------------
# Iterable
# ------------------------------------
not-an-iterable,not-a-mapping,
# -----------------------------------
# Format
# -----------------------------------
bad-indentation,mixed-indentation,unnecessary-semicolon,bad-whitespace,missing-final-newline,mixed-line-endings,
multiple-statements,trailing-newlines,trailing-whitespace,unexpected-line-ending-format,superfluous-parens,
# ------------------------------------
# Classes
# ------------------------------------
access-member-before-definition,method-hidden,assigning-non-slot,duplicate-bases,inconsistent-mro,inherit-non-class,
invalid-slots,invalid-slots-object,no-method-argument,no-self-argument,unexpected-special-method-signature,
non-iterator-returned,invalid-length-returned,protected-access,attribute-defined-outside-init,abstract-method,
bad-staticmethod-argument,non-parent-init-called,super-init-not-called,no-classmethod-decorator,
no-staticmethod-decorator,no-self-use,bad-classmethod-argument,bad-mcs-classmethod-argument,bad-mcs-method-argument,
method-check-failed,
"""
